Internal: introduce an Executor class (#8)
Problem: the `Node` class assumes that the coordinator will use the
aleph.im compute resource nodes in all cases. We wish to let the user
specify dedicated executor servers as well.

Solution: introduce an `Executor` class. If running on aleph.im, 
the API URL of the executor will be https://{node_url}/vm/{function}, 
otherwise it will default to https://{node_url}.

The result dictionaries now use executors as keys as the model is now
hashable.