Feature: allow user to change settings dynamically
Problem: most settings are configured through the `settings` global
variable. This variable is configured once when the module is imported.
This is not user-friendly when using the coordinator in library mode.
To change settings, one must import this global variable and modify it.

Solution: allow the user to pass most parameters as arguments to the
`generate_vrf()` function. If the user does not specify a value, it
defaults to the value in the settings.

In addition, the coordinator now only builds an Aleph API client once.