Feature: allow the user to specify a private key (#18)
Problem: the private key is generated automatically, leaving no control
to the user.

Solution: allow the user to specify a private key through the settings,
as is often done in other aleph.im applications.

This feature helps fixing a concurrency bug in the integration tests
caused by concurrent calls to `get_fallback_private_key()` attempting
to create a symlink at the same time. We now generate a private key in
the fixture.