Fix: pad numbers to requested size before XOR-ing them
Problem: rebuilding the random number from the random bytes in the
executor responses does not work when the generated random number is
smaller than the requested number of bytes. This occurs because the
random number is not padded before being XOR-ed with the others.

Solution: pad the random number of each executor to guarantee that it is
N bytes wide.